Move tournaments tab to components folder
