+++ app/client/api/query/useLeagues.ts
