+++ app/client/components/TournamentsTabs/index.ts
@@ -1,0 +1,3 @@
+import TournamentsTabs from "./TournamentsTabs";
+
+export default TournamentsTabs;
